rustc_sanitizers::cfi
Module typeid

Type metadata identifiers for LLVM Control Flow Integrity (CFI) and cross-language LLVM CFI support for the Rust compiler.

For more information about LLVM CFI and cross-language LLVM CFI support for the Rust compiler, see design document in the tracking issue #89653.

Modules

itanium_cxx_abi

Type metadata identifiers (using Itanium C++ ABI mangling for encoding) for LLVM Control Flow Integrity (CFI) and cross-language LLVM CFI support.

Structs

TypeIdOptions

Options for typeid_for_fnabi.

Functions

typeid_for_fnabi

Returns a type metadata identifier for the specified FnAbi.

typeid_for_instance

Returns a type metadata identifier for the specified Instance.