rustc_sanitizers::cfi::typeid::itanium_cxx_abi::encode

Function encode_predicates

fn encode_predicates<'tcx>(
    tcx: TyCtxt<'tcx>,
    predicates: &List<PolyExistentialPredicate<'tcx>>,
    dict: &mut FxHashMap<DictKey<'tcx>, usize>,
    options: TypeIdOptions,
) -> String

Encodes predicates using the Itanium C++ ABI with vendor extended type qualifiers and types for Rust types that are not used at the FFI boundary.